+++ src/persistence/gml.jl
@@ -1,4 +1,6 @@
-function _gml_read_one_graph(gs, dir)
+function readgml(io::IO)
+    gs = first(GML.parse_dict(readall(io))[:graph])
+    dir = Bool(get(gs, :directed, 0))
     nodes = [x[:id] for x in gs[:node]]
     if dir
         g = DiGraph(length(nodes))
@@ -15,39 +17,17 @@ function _gml_read_one_graph(gs, dir)
     end
     return g
 end
-function loadgml(io::IO, gname::String)
-    p = GML.parse_dict(readall(io))
-    for gs in p[:graph]
-        dir = Bool(get(gs, :directed, 0))
-        graphname = get(gs, :label, dir ? "digraph" : "graph")
-
-        (gname == graphname) && return _gml_read_one_graph(gs, dir)
-    end
-    error("Graph $gname not found")
-end
-
-function loadgml_mult(io::IO)
-    p = GML.parse_dict(readall(io))
-    graphs = Dict{String, SimpleGraph}()
-    for gs in p[:graph]
-        dir = Bool(get(gs, :directed, 0))
-        graphname = get(gs, :label, dir ? "digraph" : "graph")
-        graphs[graphname] = _gml_read_one_graph(gs, dir)
-    end
-    return graphs
-end
 
 """
-savegml(f, g, gname = "graph")
+    writegml(f, g)
 
-Writes a graph `g` with name `gname`
-to a file `f` in the
+Writes a graph `g` to a file `f` in the
 [GML](https://en.wikipedia.org/wiki/Graph_Modelling_Language) format.
 """
-function savegml(io::IO, g::SimpleGraph, gname::String = "")
+function writegml(io::IO, g::SimpleGraph)
     println(io, "graph")
     println(io, "[")
-    length(gname) > 0 && println(io, "label \"$gname\"")
+    # length(gname) > 0 && println(io, "label \"$gname\"")
     is_directed(g) && println(io, "directed 1")
     for i=1:nv(g)
         println(io,"\tnode")
@@ -66,17 +46,4 @@ function savegml(io::IO, g::SimpleGraph, gname::String = "")
     return 1
 end
 
-
-"""Writes a dictionary of (name=>graph) to a file `fn`
-
-Returns number of graphs written.
-"""
-function savegml_mult(io::IO, graphs::Dict)
-    ng = 0
-    for (gname, g) in graphs
-        ng += savegml(io, g, gname)
-    end
-    return ng
-end
-
-filemap[:gml] = (loadgml, loadgml_mult, savegml, savegml_mult)
+filemap[:gml] = (readgml, writegml)