+++ src/persistence/graphml.jl
@@ -25,7 +25,7 @@ function _graphml_read_one_graph(e::XMLElement, isdirected::Bool)
     return g
 end
 
-function loadgraphml(io::IO, gname::String)
+function readgraphml(io::IO)
     xdoc = parse_string(readall(io))
     xroot = root(xdoc)  # an instance of XMLElement
     name(xroot) == "graphml" || error("Not a GraphML file")
@@ -38,80 +38,48 @@ function loadgraphml(io::IO, gname::String)
                 edgedefault = attribute(e, "edgedefault")
                 isdir = edgedefault=="directed" ? true :
                              edgedefault=="undirected" ? false : error("Unknown value of edgedefault: $edgedefault")
-                if has_attribute(e, "id")
-                    graphname = attribute(e, "id")
-                else
-                    graphname =  isdir ? "digraph" : "graph"
-                end
-                gname == graphname && return _graphml_read_one_graph(e, isdir)
+                # if has_attribute(e, "id")
+                #     graphname = attribute(e, "id")
+                # else
+                #     graphname =  isdir ? "digraph" : "graph"
+                # end
+                return _graphml_read_one_graph(e, isdir)
             else
                 warn("Skipping unknown XML element '$(name(e))'")
             end
         end
     end
-    error("Graph $gname not found")
+    error("Graph not found")
 end
 
-function loadgraphml_mult(io::IO)
-    xdoc = parse_string(readall(io))
-    xroot = root(xdoc)  # an instance of XMLElement
-    name(xroot) == "graphml" || error("Not a GraphML file")
 
-    # traverse all its child nodes and print element names
-    graphs = Dict{String, SimpleGraph}()
-    for c in child_nodes(xroot)  # c is an instance of XMLNode
-        if is_elementnode(c)
-            e = XMLElement(c)  # this makes an XMLElement instance
-            if name(e) == "graph"
-                edgedefault = attribute(e, "edgedefault")
-                isdir = edgedefault=="directed" ? true :
-                             edgedefault=="undirected" ? false : error("Unknown value of edgedefault: $edgedefault")
-                if has_attribute(e, "id")
-                    graphname = attribute(e, "id")
-                else
-                    graphname =  isdir ? "digraph" : "graph"
-                end
-                graphs[graphname] =  _graphml_read_one_graph(e, isdir)
-            else
-                warn("Skipping unknown XML element '$(name(e))'")
-            end
-        end
-    end
-    return graphs
-end
-
-function savegraphml_mult(io::IO, graphs::Dict)
+function writegraphml(io::IO, g::SimpleGraph)
     xdoc = XMLDocument()
     xroot = create_root(xdoc, "graphml")
     set_attribute(xroot,"xmlns","http://graphml.graphdrawing.org/xmlns")
     set_attribute(xroot,"xmlns:xsi","http://www.w3.org/2001/XMLSchema-instance")
     set_attribute(xroot,"xsi:schemaLocation","http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd")
 
-    for (gname, g) in graphs
-        xg = new_child(xroot, "graph")
-        set_attribute(xg,"id",gname)
-        strdir = is_directed(g) ? "directed" : "undirected"
-        set_attribute(xg,"edgedefault",strdir)
+    xg = new_child(xroot, "graph")
+    # set_attribute(xg,"id",gname)
+    strdir = is_directed(g) ? "directed" : "undirected"
+    set_attribute(xg,"edgedefault",strdir)
 
-        for i=1:nv(g)
-            xv = new_child(xg, "node")
-            set_attribute(xv,"id","n$(i-1)")
-        end
+    for i=1:nv(g)
+        xv = new_child(xg, "node")
+        set_attribute(xv,"id","n$(i-1)")
+    end
 
-        m = 0
-        for e in edges(g)
-            xe = new_child(xg, "edge")
-            set_attribute(xe,"id","e$m")
-            set_attribute(xe,"source","n$(src(e)-1)")
-            set_attribute(xe,"target","n$(dst(e)-1)")
-            m += 1
-        end
+    m = 0
+    for e in edges(g)
+        xe = new_child(xg, "edge")
+        set_attribute(xe,"id","e$m")
+        set_attribute(xe,"source","n$(src(e)-1)")
+        set_attribute(xe,"target","n$(dst(e)-1)")
+        m += 1
     end
     show(io, xdoc)
-    return length(graphs)
+    return 1
 end
 
-savegraphml(io::IO, g::SimpleGraph, gname::String) =
-    savegraphml_mult(io, Dict(gname=>g))
-
-filemap[:graphml] = (loadgraphml, loadgraphml_mult, savegraphml, savegraphml_mult)
+filemap[:graphml] = (readgraphml, writegraphml)