+++ src/persistence/net.jl
@@ -4,7 +4,7 @@ Writes a graph `g` to a file `f` in the [Pajek
 NET](http://gephi.github.io/users/supported-graph-formats/pajek-net-format/) format.
 Returns 1 (number of graphs written).
 """
-function savenet(f::IO, g::SimpleGraph, gname::String = "g")
+function writepajek(f::IO, g::SimpleGraph)
     println(f, "*Vertices $(nv(g))")
     # write edges
     if is_directed(g)
@@ -23,7 +23,7 @@ end
  NET](http://gephi.github.io/users/supported-graph-formats/pajek-net-format/) format.
  Returns 1 (number of graphs written).
 """
-function loadnet(f::IO, gname::String = "g")
+function readpajek(f::IO)
     line =readline(f)
     # skip comments
     while startswith(line, "%")
@@ -60,6 +60,4 @@ function loadnet(f::IO, gname::String = "g")
     return g
 end
 
-loadnet_mult(io::IO) = Dict("g" => loadnet(io))
-
-filemap[:net] = (loadnet, loadnet_mult, savenet, NI)
+filemap[:NET] = (readpajek, writepajek)